Create 3) Inferential Statistics and Hypothesis Testing.R
A. Package ‘MASS’ which provides a description of the datasets available in the MASS package 

Then, complete the following analysis of the identified data from the library. 

B. One-sample t-test: 
Use the “chem” dataset to answer the question, “is the flour production company producing whole meal flour with greater than 1 part per million copper in it?” 

C. Two-sample t-test: 
Use the “cats” dataset to answer the question, “do male and female cat samples have the same body weight?” 

Hint: one way to get separate vectors for male and female cat body weight is to use the subset function as follows: “male <-subset(cats, subset=(cats$Sex=M))” 

D. Paired t-test: 
Use the “shoes” dataset to answer the question, “did material A wear better than material B?” 

E. Test of equal or given proportions: 
Use the “bacteria” data set to answer the question, “did the drug treatment have a significant effect of the presence of the bacteria compared with the placebo?” 

F. F-test: 
Use the “cats” data set to test for the variance of the body weight in male and female cats.